:construction: `ls-remote` cmd supports windows
